&&解决动画系统偶尔引起崩溃的问题 &&解决setScaleX 偶尔引起位置改变的问题

+++ CrossApp/view/CAAlertView.cpp
@@ -248,7 +248,7 @@ void CAAlertView::adjustButtonView() {
 
 	float alertViewButtonHeight = 88;
 
-	int btnCount = m_vAllBtn.size();
+	size_t btnCount = m_vAllBtn.size();
 
 	CC_RETURN_IF(btnCount <= 0);
